First attempt at running Imogen on GPU - THIS IS BROKEN DO NOT USE IT EXCEPT TO FIX

+++ save/SaveManager.m
@@ -248,7 +248,7 @@ classdef SaveManager < handle
 			if iscodistributed(array)
 				result = squeeze( gather(array(i{:}),1) );
 			else
-				result = squeeze( array(i{:}) );
+				result = double(squeeze( array(i{:}) ));
 			end
 		end
